Rename: 디렉토리 이전
- `/v1/*/domain/enums` -> `/v1/*/domain/code`
