+++ src/main/java/com/vigilante/retriever/v1/bookmark/domain/code/BookmarkErrorCode.java
@@ -1,4 +1,4 @@
-package com.vigilante.retriever.v1.post.domain.enums;
+package com.vigilante.retriever.v1.bookmark.domain.code;
 
 import com.vigilante.retriever.common.domain.code.BaseCode;
 
@@ -7,8 +7,8 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum PostErrorCode implements BaseCode {
-	POST_NOT_FOUND("POST-4041", "해당하는 게시글을 찾을 수 없습니다.");
+public enum BookmarkErrorCode implements BaseCode {
+	BOOKMARK_NOT_FOUND("BOOKMARK-4041", "해당하는 채널 북마크를 찾을 수 없습니다.");
 
 	private final String code;
 	private final String message;
