+++ src/main/java/com/vigilante/retriever/v1/channel/domain/exception/ChannelDataNotFoundException.java
@@ -1,6 +1,6 @@
 package com.vigilante.retriever.v1.channel.domain.exception;
 
-import static com.vigilante.retriever.v1.channel.domain.enums.ChannelDataErrorCode.*;
+import static com.vigilante.retriever.v1.channel.domain.code.ChannelDataErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
