+++ src/main/java/com/vigilante/retriever/v1/channel/domain/exception/ChannelInfoNotFoundException.java
@@ -1,6 +1,6 @@
 package com.vigilante.retriever.v1.channel.domain.exception;
 
-import static com.vigilante.retriever.v1.channel.domain.enums.ChannelInfoErrorCode.*;
+import static com.vigilante.retriever.v1.channel.domain.code.ChannelInfoErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
