+++ src/main/java/com/vigilante/retriever/v1/channel/domain/exception/ChannelNotFoundException.java
@@ -1,6 +1,6 @@
 package com.vigilante.retriever.v1.channel.domain.exception;
 
-import static com.vigilante.retriever.v1.channel.domain.enums.ChannelErrorCode.*;
+import static com.vigilante.retriever.v1.channel.domain.code.ChannelErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
