+++ src/main/java/com/vigilante/retriever/v1/channel/domain/exception/ChannelSimilarityNotFoundException.java
@@ -1,6 +1,6 @@
 package com.vigilante.retriever.v1.channel.domain.exception;
 
-import static com.vigilante.retriever.v1.channel.domain.enums.ChannelSimilarityErrorCode.*;
+import static com.vigilante.retriever.v1.channel.domain.code.ChannelSimilarityErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
