+++ src/main/java/com/vigilante/retriever/v1/chatbot/domain/code/ChatBotErrorCode.java
@@ -1,4 +1,4 @@
-package com.vigilante.retriever.v1.channel.domain.enums;
+package com.vigilante.retriever.v1.chatbot.domain.code;
 
 import com.vigilante.retriever.common.domain.code.BaseCode;
 
@@ -7,8 +7,8 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum ChannelErrorCode implements BaseCode {
-	CHANNEL_NOT_FOUND("CHANNEL-4041", "해당하는 채널을 찾을 수 없습니다.");
+public enum ChatBotErrorCode implements BaseCode {
+	CHAT_BOT_NOT_FOUND("CHAT_BOT-4041", "해당하는 챗봇을 찾을 수 없습니다.");
 
 	private final String code;
 	private final String message;
