+++ src/main/java/com/vigilante/retriever/v1/drug/domain/code/DrugErrorCode.java
@@ -1,4 +1,4 @@
-package com.vigilante.retriever.v1.bookmark.domain.enums;
+package com.vigilante.retriever.v1.drug.domain.code;
 
 import com.vigilante.retriever.common.domain.code.BaseCode;
 
@@ -7,8 +7,8 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum BookmarkErrorCode implements BaseCode {
-	BOOKMARK_NOT_FOUND("BOOKMARK-4041", "해당하는 채널 북마크를 찾을 수 없습니다.");
+public enum DrugErrorCode implements BaseCode {
+	DRUG_NOT_FOUND("DRUG-4041", "해당하는 챗봇을 찾을 수 없습니다.");
 
 	private final String code;
 	private final String message;
