+++ src/main/java/com/vigilante/retriever/v1/report/domain/code/ReportErrorCode.java
@@ -1,4 +1,4 @@
-package com.vigilante.retriever.v1.user.domain.enums;
+package com.vigilante.retriever.v1.report.domain.code;
 
 import com.vigilante.retriever.common.domain.code.BaseCode;
 
@@ -7,8 +7,8 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum RefreshTokenErrorCode implements BaseCode {
-	REFRESH_TOKEN_NOT_FOUND("TOKEN-4041", "리프레쉬 토큰이 존재하지 않습니다");
+public enum ReportErrorCode implements BaseCode {
+	REPORT_NOT_FOUND("REPORT-4041", "해당하는 Ai 채널 리포트를 찾을 수 없습니다.");
 
 	private final String code;
 	private final String message;
