+++ src/main/java/com/vigilante/retriever/v1/report/domain/exception/ReportNotFoundException.java
@@ -1,6 +1,6 @@
 package com.vigilante.retriever.v1.report.domain.exception;
 
-import static com.vigilante.retriever.v1.report.domain.enums.ReportErrorCode.*;
+import static com.vigilante.retriever.v1.report.domain.code.ReportErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
