+++ src/main/java/com/vigilante/retriever/v1/user/domain/exception/UserNotFoundException.java
@@ -1,6 +1,6 @@
 package com.vigilante.retriever.v1.user.domain.exception;
 
-import static com.vigilante.retriever.v1.user.domain.enums.UserErrorCode.*;
+import static com.vigilante.retriever.v1.user.domain.code.UserErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
